Fixed reusing dbh when sth is still exists.
When we using async operations, and got db and results object - we can pass result object to subfunction. After that result object continues to exists, and db - not, so DESTROY on db called and dbh returns to queue.
See https://github.com/kraih/mojo-pg/issues/36.

So we can store db object to results object to prevent this behavior.